api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.6
python-dotenv==1.0.0
python-multipart==0.0.19
langchain==0.3.27
langchain-google-genai==2.1.8
piper-tts==1.3.0
openai-whisper==20250625